examprep-gh/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pdfplumber
supabase
openai
python-jose[cryptography]
passlib[bcrypt]
httpx
